Prep for C++ conversion
Fixed all warnings and errors that occurred after I renamed the .c files to
their .cpp equivalents.
As a side effect, this fixed up some code in transTable.c that had some
corner cases in the size calculations, and fixed a castling-byte bug in the
console-UI edit-board code.
Also, it makes the code run very slightly faster w/gcc.
Also, the code runs slightly faster w/g++ than w/gcc, so my fears regarding
overhead for exception support would appear to be unfounded.
In any case, cleaning up the warnings and errors was a good thing.

+++ pv.c
@@ -161,7 +161,7 @@ int PvBuildMoveString(PvT *pv, char *dstStr, int dstLen,
 			 i == 0 ? "" : " ",
 			 sanStr);
 	myStrLen = myStr - myStrSpace;
-	assert(myStrLen < sizeof(myStrSpace));
+	assert(myStrLen < (int) sizeof(myStrSpace));
 	if (myStrLen > dstLen)
 	{
 	    // We wrote too much information.  Chop the last move off.
